Improve ble decryption
