+++ transport_plugins/bled112/iotile_transport_bled112/bled112.py
@@ -7,8 +7,10 @@ import threading
 import logging
 import datetime
 import copy
+import struct
 import serial
 import serial.tools.list_ports
+from Crypto.Cipher import AES
 from iotile_transport_bled112 import bgapi_structures
 from iotile.core.dev.config import ConfigManager
 from iotile.core.utilities.packed import unpack
@@ -19,7 +21,6 @@ from .bled112_cmd import BLED112CommandProcessor
 from .tilebus import TileBusService, TileBusStreamingCharacteristic, TileBusTracingCharacteristic, TileBusHighSpeedCharacteristic
 from .async_packet import AsyncPacketBuffer
 from .utilities import open_bled112
-from .security import verify_payload, generate_nonce
 from iotile.core.hw.auth.auth_provider import AuthProvider
 from iotile.core.hw.auth.auth_chain import ChainedAuthProvider
 
@@ -34,6 +35,22 @@ def packet_length(header):
 
     return (highbits << 8) | lowbits
 
+
+def generate_nonce(device_uuid, timestamp, low_reboots, high_reboots, counter_packed):
+    return struct.pack("<LLHBBB", device_uuid, timestamp, low_reboots, high_reboots, counter_packed, 0)
+
+
+def decrypt_payload(key, message, nonce):
+    aad = message[0:14]
+    body = message[14:20]
+    mac = message[20:]
+
+    cipher = AES.new(key, AES.MODE_CCM, nonce, mac_len=4)
+    cipher.update(aad)
+    decrypted_data = cipher.decrypt(body)
+    cipher.verify(mac)
+
+    return decrypted_data
 
 class BLED112Adapter(DeviceAdapter):
     """Callback based BLED112 wrapper supporting multiple simultaneous connections.
@@ -636,32 +653,6 @@ class BLED112Adapter(DeviceAdapter):
         broadcast_multiplex = counter_packed >> 5
         broadcast_toggle = broadcast_stream_packed >> 15
         broadcast_stream = broadcast_stream_packed & ((1 << 15) - 1)
-        is_encrypted = bool(flags & (1 << 3))
-        is_device_key = bool(flags & (1 << 4))
-
-        key_type = AuthProvider.NoKey
-        if is_encrypted:
-            if is_device_key:
-                key_type = AuthProvider.DeviceKey
-            else:
-                key_type = AuthProvider.UserKey
-
-        if is_encrypted:
-            try:
-                key = self._key_provider.get_rotated_key(key_type, device_id,
-                    reboot_counter=reboots,
-                    rotation_interval_power=EPHEMERAL_KEY_CYCLE_POWER,
-                    current_timestamp=timestamp)
-                nonce = generate_nonce(device_id, timestamp, reboot_low, reboot_high_packed, counter_packed)
-
-                if not verify_payload(key, data[7:], nonce):
-                    print(reboots, timestamp)
-                    self._logger.warning("Advertisement packet is not verified")
-                    return
-            except NotFoundError:
-                self._logger.warning("Key is not found")
-                return
-
 
         # Flags for version 2 are:
         #   bit 0: Has pending data to stream
@@ -672,14 +663,50 @@ class BLED112Adapter(DeviceAdapter):
         #   bit 5: Encryption key is user key
         #   bit 6: broadcast data is time synchronized to avoid leaking
         #   information about when it changes
+        is_pending_data = bool(flags & (1 << 0))
+        is_low_voltage = bool(flags & (1 << 1))
+        is_user_connected = bool(flags & (1 << 2))
+        is_encrypted = bool(flags & (1 << 3))
+        is_device_key = bool(flags & (1 << 4))
+        is_user_key = bool(flags & (1 << 5))
 
         self._device_scan_counts.setdefault(device_id, {'v1': 0, 'v2': 0})['v2'] += 1
 
+        key_type = AuthProvider.NoKey
+        if is_encrypted:
+            if is_device_key:
+                key_type = AuthProvider.DeviceKey
+            elif is_user_key:
+                key_type = AuthProvider.UserKey
+
+        if is_encrypted and (is_device_key or is_user_key):
+            try:
+                key = self._key_provider.get_rotated_key(key_type, device_id,
+                    reboot_counter=reboots,
+                    rotation_interval_power=EPHEMERAL_KEY_CYCLE_POWER,
+                    current_timestamp=timestamp)
+            except NotFoundError:
+                self._logger.warning("Key type {} is not found".format(key_type))
+                return
+
+            nonce = generate_nonce(device_id, timestamp, reboot_low, reboot_high_packed, counter_packed)
+
+            try:
+                decrypted_data = decrypt_payload(key, data[7:], nonce)
+            except ValueError:
+                print(hex(device_id), hex(reboots), hex(timestamp), key.hex())
+                self._logger.warning("Advertisement packet is not verified")
+                return
+
+            broadcast_stream_packed, broadcast_value = unpack("<HL", decrypted_data)
+            broadcast_toggle = broadcast_stream_packed >> 15
+            broadcast_stream = broadcast_stream_packed & ((1 << 15) - 1)
+
         info = {'connection_string': sender,
                 'uuid': device_id,
-                'pending_data': bool(flags & (1 << 0)),
-                'low_voltage': bool(flags & (1 << 1)),
-                'user_connected': bool(flags & (1 << 2)),
+                'pending_data': is_pending_data,
+                'low_voltage': is_low_voltage,
+                'user_connected': is_user_connected,
                 'signal_strength': rssi,
                 'reboot_counter': reboots,
                 'sequence': counter,
